Unify transfer tests
